feat: select a random art with every new message fixes #137
