+++ arts.js
@@ -1,0 +1,40 @@
+var bunnySay = require('sign-bunny');
+var yosay = require('yosay');
+
+var llamaSay = function(text) {
+  return `
+    ${text}
+
+    ∩∩
+　（･ω･）
+　　│ │
+　　│ └─┐○
+　  ヽ　　　丿
+　　 　∥￣∥`;
+}
+
+var catSay = function(text) {
+  return `
+      ${text}
+
+      ♪ ガンバレ! ♪
+  ミ ゛ミ ∧＿∧ ミ゛ミ
+  ミ ミ ( ・∀・ )ミ゛ミ
+   ゛゛ ＼　　　／゛゛
+   　　 　i⌒ヽ ｜
+  　　 　 (＿) ノ
+   　　　　　 ∪`
+    ;
+}
+
+function getCustomArt(art) {
+	switch (art) {
+		case 'bunny' : return bunnySay(textToSay);
+		case 'llama' : return llamaSay(textToSay);
+		case 'cat'   : return catSay(textToSay);
+		case 'yeoman': return yosay(textToSay);
+		default : return null;
+	}
+}
+
+module.exports = getCustomArt;
