+++ package.json
@@ -13,15 +13,15 @@
   },
   "repository": "https://github.com/notwaldorf/tiny-care-terminal.git",
   "dependencies": {
-    "ansi-art": "^1.2.4",
+    "ansi-art": "^1.3.0",
     "blessed": "^0.1.81",
     "blessed-contrib": "^4.8.5",
     "chalk": "^2.4.1",
+    "cowsay": "^1.4.0",
     "expand-home-dir": "^0.0.3",
-    "cowsay": "^1.2.1",
     "git-user-name": "^2.0.0",
-    "gitlog": "^3.1.2",
     "git-utils": "^5.4.0",
+    "gitlog": "^3.1.2",
     "is-git": "^1.0.0",
     "node-notifier": "^5.2.1",
     "resolve-dir": "^1.0.1",
