+++ sample.env
@@ -4,8 +4,8 @@ export TTC_BOTS='tinycarebot,selfcare_bot,magicrealismbot'
 
 # Use this to have a different animal say a message in the big box.
 # See README for details about supported formats.
-# Use e.g 'bunny.cow', 'llama' or 'bunny'
-export TTC_SAY_BOX='panda.ansi'
+# Use e.g 'bunny.cow', 'panda.ansi', 'llama' or 'RANDOM'.
+export TTC_SAY_BOX='RANDOM'
 
 # List of folders to look into for `git` commits, comma separated.
 export TTC_REPOS='~/Work/Code,~/Fun/Code'
